requirements.txt
browser-use==1.4.2
pandas>=2.0.0
pytrends==5.0.0
beautifulsoup4==4.12.0
requests==2.31.0
pyyaml==6.0.1